refactor: function that combines genebe and clinvar results is moved to geneBe_utils function (#42)
